pyproject.toml
[tool.poetry]
name = "clearedge"
version = "0.1.14"
description = ""
authors = ["Clearedge"]
readme = "README.md"

[tool.poetry.scripts]
install-deps = "install_deps:main"

[tool.poetry.dependencies]
python = ">=3.9,<3.12"
requests = "^2.31.0"
validators = "^0.22.0"
rapidocr-onnxruntime = "^1.3.15"
ultralytics = ">=8.0.225,<8.1.0"
python-doctr = {extras = ["torch"], version = "^0.8.1"}
huggingface-hub = "^0.22.0"
pytesseract = "^0.3.10"
pdf2image = "^1.17.0"
rapid-table = "^0.1.3"
tqdm = "^4.66.2"
pymupdf = "^1.24.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
